Updated Apple Pay icon
The previous SVG icon did not match Apple's official EPS logo download.  The official icon is slightly taller and has a slightly thicker type and stroke.

Additionally, the previous version of the icon used unusual combinations of masks and paths that were problematic if you applied CSS fills to the path.