Remove test expecting the exception described in issue #524 and removed in PR #532. Remove CSS code intended to cause that exception to be thrown.
With PR #532 I changed slate-cssvar-loader to allow deploying when it encountered CSS variables without corresponding Liquid variables, issuing a warning via the console instead of throwing an unhandled exception. The commit in the PR didn't remove that test. This commit removes that test and the CSS code that test depended on.