Update core.js
Now sections has ability to have separately placed json files to be used as schema.
The idea is:
We have section files in  `/src/sections/`  like `/src/sections/section_name.liquid` and  to simplify editing of section's schema which is json
We now can create just `/src/sections/section_name.schema.json` file and it'll be injected to same named section liquid file upon build process and be placed  wrapped with right `{% schema %}` tags in `/dist/sections/section_name.liquid`
Same logic can be implemented for Scripts and Styles. But i consider Slate's approach is to separate JS and CSS\SCSS out of sections to be built  with webpack instead of having them inlined in sections, even Shopify supports it.