Update Create-a-self-signed-SSL-certificate.md
Update shell script to create SSL certificate for all IPs found by `hostname`. This improves accuracy when running ssl-check. For instance, if you have a docker network interface, then sometimes the first IP will actually be the docker IP. The ssl-check function would originally fail because of that. This change resolves that failure.